aider: Cast the result of `get_current_timestamp() + 60*60*24*7` to `usize` in the `expires` field of the `Claims` struct.
# Aider chat conversation:

USER: I ran this command:

 cargo run

And got this output:

   Compiling agenthost v0.1.0 (/home/runvnc/agenthost)
error[E0308]: mismatched types
  --> src/jwt_util/mod.rs:25:18
   |
25 |         expires: get_current_timestamp() + 60*60*24*7 as u64,
   |                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ expected `usize`, found `u64`

For more information about this error, try `rustc --explain E0308`.
error: could not compile `agenthost` (bin "agenthost") due to previous error

ASSISTANT: The error indicates that the `expires` field in the `Claims` struct is expected to be of type `usize`, but the expression `get_current_timestamp() + 60*60*24*7 as u64` results in a `u64`. To fix this, you need to cast the result of the entire expression to `usize` since Rust does not implicitly convert between numeric types.

Here's the change you need to make:

```rust
src/jwt_util/mod.rs
<<<<<<< SEARCH
    expires: get_current_timestamp() + 60*60*24*7 as u64,
=======
    expires: (get_current_timestamp() + 60*60*24*7) as usize,
>>>>>>> REPLACE
```

This change casts the result of `get_current_timestamp() + 60*60*24*7` to `usize` to match the expected type of the `expires` field.

+++ src/jwt_util/mod.rs
@@ -22,7 +22,7 @@ pub struct Claims {
 pub fn create_token(user_id: &str) -> Result<String, (StatusCode, &'static str)> {
     let my_claims = Claims {
         username: user_id.to_owned(),
-        expires: get_current_timestamp() + 60*60*24*7 as u64,
+        expires: (get_current_timestamp() + 60*60*24*7) as usize,
     };
     let encoding_key = jwt::EncodingKey::from_secret(CODE.as_ref());
     let token = encode(&Header::default(), &my_claims, &encoding_key)
